+++ src/main/java/com/telepathicgrunt/repurposedstructures/misc/EndRemasteredDedicatedLoot.java
@@ -6,7 +6,6 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.LootContext;
-import net.minecraft.world.level.storage.loot.LootTable;
 
 import java.util.HashMap;
 import java.util.List;